Added support for std::complex buffers to interpolator classes and FloatVectorOperations
Converted LagrangeInterpolator and CatmullRomInterpolator to templated classes making it possible to specify a double or std::complex sample as SampleType. Added interfaces for std::complex values to FloatVectorOperations for all useful cases.
