+++ modules/juce_audio_basics/juce_audio_basics.cpp
@@ -58,8 +58,6 @@
 #include "buffers/juce_AudioChannelSet.cpp"
 #include "buffers/juce_AudioProcessLoadMeasurer.cpp"
 #include "utilities/juce_IIRFilter.cpp"
-#include "utilities/juce_LagrangeInterpolator.cpp"
-#include "utilities/juce_CatmullRomInterpolator.cpp"
 #include "utilities/juce_SmoothedValue.cpp"
 #include "midi/juce_MidiBuffer.cpp"
 #include "midi/juce_MidiFile.cpp"
